# pubapi/openapi/v1beta.yml
openapi: 3.1.0
info:
  title: Marble decision engine API
  description: Public API on decisions and data ingestion.
  version: 1.0.0
...
servers:
  - url: https://api.checkmarble.com/v1beta
...
paths:
...
  /cases:
    get:
      operationId: listCases
      tags:
        - Cases
      security:
        - BearerTokenAuth: []
        - ApiKeyAuth: []
      summary: List cases
      description: |
        List cases according to a filter set.

        This endpoint will list all cases matching the filter, sorted by creation date, in the requested order. It will not reorder "waiting for action" or hide snoozed cases.
      parameters:
        - name: inbox_id
          description: |
            Only list cases in the specified inboxes.

            This parameter may be repeated as necessary.
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          description: |
            Filter cases by their investigation status.

            This parameter may be repeated as necessary.
          in: query
          schema:
            type: string
            enum: [pending, investigating, closed]
        - name: assigned_to
          description: List cases assigned to the specified user
          in: query
          schema:
            type: string
            format: uuid
        - name: start
          description: Filter cases created on and after this date
          in: query
          schema:
            type: string
            format: date-time
        - name: end
          description: Filter cases created strictly before this date
          in: query
          schema:
            type: string
            format: date-time
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/order"
        - $ref: "#/components/parameters/after"
      responses:
        "200":
          description: List of cases
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - $ref: "#/components/schemas/BasePagination"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Case"
        "400":
          $ref: "#/components/responses/400"

    post:
      operationId: createCase
      tags:
        - Cases
      security:
        - BearerTokenAuth: []
        - ApiKeyAuth: []
      summary: Create new case
      description: |
        Create a new case.

        This endpoint enforces the following preconditions:

          - All decisions exist and are not already assigned to a case
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - inbox_id
                - name
              properties:
                inbox_id:
                  description: ID of the inbox to create the case in
                  type: string
                  format: uuid
                name:
                  description: Name of the case
                  type: string
                assignee_email:
                  description: Email address of the assignee
                  type: string
                  format: email
                decisions:
                  description: List of decision_ids to attach to this case. Must not already be assigned to a case.
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        "200":
          description: New case
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Case"
        "400":
          $ref: "#/components/responses/400"
        "422":
          $ref: "#/components/responses/422"
  /cases/{caseId}:
    get:
      operationId: getCase
      tags:
        - Cases
      security:
        - BearerTokenAuth: []
        - ApiKeyAuth: []
      summary: Get case
      description: |
        Get case by its ID.
      parameters:
        - name: caseId
          required: true
          in: path
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Case
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Case"
        "400":
          $ref: "#/components/responses/400"

    patch:
      operationId: updateCase
      tags:
        - Cases
      security:
        - BearerTokenAuth: []
        - ApiKeyAuth: []
      summary: Update case
      description: |
        Update a case.

        Some actions will be available through specific endpoints and will not be updateable from here.
      parameters:
        - name: caseId
          required: true
          in: path
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                inbox:
                  description: ID of the inbox to move the case in
                  type: string
                  format: uuid
                name:
                  description: Name of the case
                  type: string
                outcome:
                  description: Outcome of the case
                  type: string
                  enum: [unset confirmed_risk valuable_alert false_positive]

      responses:
        "200":
          description: Case
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Case"
        "400":
          $ref: "#/components/responses/400"
        "422":
          $ref: "#/components/responses/422"
...
  /cases/{caseId}/escalate:
    post:
      operationId: escalateCase
      tags:
        - Cases
      security:
        - BearerTokenAuth: []
        - ApiKeyAuth: []
      summary: Escalate case
      description: |
        Escalate a case, moving it to the configured escalation inbox and marking it as waiting for action.

        This endpoint enforces the following preconditions:

          - The source inbox is configured to have an escalation inbox
      parameters:
        - name: caseId
          required: true
          in: path
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Case
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Case"
        "400":
          $ref: "#/components/responses/400"
        "422":
          $ref: "#/components/responses/422"
  /cases/{caseId}/comments:
    get:
      operationId: listCaseComments
      tags:
        - Cases
      security:
        - BearerTokenAuth: []
        - ApiKeyAuth: []
      summary: List case comments
      description: |
        List comments that were left on the given case.
      parameters:
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/order"
        - $ref: "#/components/parameters/after"
        - name: caseId
          required: true
          in: path
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of comments
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - $ref: "#/components/schemas/BasePagination"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/CaseComment"
        "400":
          $ref: "#/components/responses/400"
        "413":
          $ref: "#/components/responses/413"

    post:
      operationId: createCaseComment
      tags:
        - Cases
      security:
        - BearerTokenAuth: []
        - ApiKeyAuth: []
      summary: Create new case comment
      description: |
        Write a new comment on the provided case.
      parameters:
        - name: caseId
          required: true
          in: path
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - comment
              properties:
                comment:
                  description: |
                    Body of the comment.

                    You can format the comment with markdown markup. The comment must be up to 2048 characters.
                  type: string
                  maxLength: 2048
      responses:
        "201":
          description: Comment created
        "400":
          $ref: "#/components/responses/400"
        "422":
          $ref: "#/components/responses/422"
...
  /cases/{caseId}/reviews:
    get:
      operationId: listCaseReviews
      tags:
        - Cases
      security:
        - BearerTokenAuth: []
        - ApiKeyAuth: []
      summary: List AI case reviews
      description: |
        List AI reviews generated for the given case, ordered by creation date.
        The review content is not included — use the get review by ID endpoint to fetch the full content.
      parameters:
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/order"
        - $ref: "#/components/parameters/after"
        - name: caseId
          required: true
          in: path
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of AI case reviews
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - $ref: "#/components/schemas/BasePagination"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/CaseReview"
        "400":
          $ref: "#/components/responses/400"
  /cases/{caseId}/reviews/{reviewId}:
    get:
      operationId: getCaseReview
      tags:
        - Cases
      security:
        - BearerTokenAuth: []
        - ApiKeyAuth: []
      summary: Get AI case review
      description: |
        Get an AI case review by its ID. The full review content is included when the status is "completed".
      parameters:
        - name: caseId
          required: true
          in: path
          schema:
            type: string
            format: uuid
        - name: reviewId
          required: true
          in: path
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: AI case review
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/CaseReviewDetail"
        "400":
          $ref: "#/components/responses/400"
        "404":
          $ref: "#/components/responses/404"
...
components:
  securitySchemes:
    BearerTokenAuth:
      type: http
      scheme: bearer
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
  responses:
...
    "400":
      description: Provided parameters or payload is malformed or invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    "404":
      description: The requested resource does not exist
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    "413":
      description: The request body exceeds maximum allowed size.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    "422":
      description: The requested action is not possible on the requested resource
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  parameters:
    limit:
      name: limit
      description: How many item to return on each page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
    order:
      name: order
      description: Sort direction for items
      in: query
      schema:
        type: string
        enum: ["ASC", "DESC"]
    after:
      name: after
      description: |
        Value to start iterating from.

        Should be a value returned in the `pagination.next_page_id` field of a response.
      in: query
      schema:
        type: string
  schemas:
    BaseResponse:
      title: Base response
      type: object
      required:
        - data
      properties:
        data:
          description: Requested objects from the API
    BasePagination:
      type: object
      properties:
        pagination:
          title: Pagination information
          type: object
          required:
            - has_more
          properties:
            has_more:
              description: Whether there are more elements to iteration through
              type: boolean
            next_page_id:
              description: Value to use in pagination filter to request for the next page
              type: string
...